% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_dna.R
\name{dist_dna}
\alias{dist_dna}
\title{Compute pairwise genetic distances}
\usage{
dist_dna(x)
}
\arguments{
\item{x}{A list of integer vectors, each of which represents the genome of
a pathogen as positions of mutations.}
}
\description{
This function Compute pairwise genetic distances from vectors of mutations as
generated in the simulations. The computed distance is the Hamming distance,
i.e. the number of nucleotide differences between sequences.
}
\examples{
dist_dna(list(integer(0), integer(0), 1, 1))
dist_dna(list(integer(0), integer(0), 1, 1, 1:2, 1:3, 1:4, 11:20))

}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
