% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_case.R, R/print.case.R
\name{new_case}
\alias{new_case}
\alias{print.case}
\title{Create a new case}
\usage{
new_case(infector = NULL, date = NULL, ..., config = new_config(...))

\method{print}{case}(x, ...)
}
\arguments{
\item{infector}{An optional infector; if provided, it must be a list with the
following slots: 'location', 'dna'}

\item{date}{The date at which the new case is infected; if the case is
imported and this is not provided, defaults to 0.}

\item{...}{further arguments passed to \code{\link{new_config}}}

\item{config}{optionally, a config returned by \code{\link{new_config}}}

\item{x}{A 'case' object, as returned by \code{\link{new_case}}.}
}
\description{
This function creates a new case, with or without a known infector. If it is
missing (NULL, default), the case is imported according to parameters in
\code{config}.
}
\details{
It calls upon \code{\link{new_location}} and \code{\link{new_dna}} to
determine locations and DNA sequences.
}
\examples{

## imported cases
A <- new_case()
B <- new_case()
A
B

## transmission chain A -> C -> D
C <- new_case(A, date = 3)
D <- new_case(C, date = 10)

A
B
C
D
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
