% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_config.R
\name{new_config}
\alias{new_config}
\title{Create a simulation configuration}
\usage{
new_config(...)
}
\arguments{
\item{...}{named simulation settings - see details}
}
\description{
This function creates a list storing simulation settings - see details.
}
\details{
Simulation settings include:
\itemize{
  \item \code{x_min}: lower bound of spatial coordinates (x-axis)
  \item \code{x_max}: upper bound of spatial coordinates (x-axis)
  \item \code{y_min}: lower bound of spatial coordinates (y-axis)
  \item \code{y_max}: upper bound of spatial coordinates (y-axis)
  \item \code{sd_spatial}: standard deviation of the Normal kernel use in
    spatial dispersion
  \item \code{genome_length}: number of nucleotides in the pathogen genome
  \item \code{mutation_rate}: per nucleotide and day
  \item \code{separation_lineages}: number of days to ancestral lineage for
    newly introduced pathogens 
}
}
\examples{

new_config()
new_config(genome_length = 100, xmax = 10)

}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
