% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_dna.R
\name{new_dna}
\alias{new_dna}
\title{Create the location of a new case}
\usage{
new_dna(dna = NULL, generation_time = NULL, ..., config = new_config(...))
}
\arguments{
\item{dna}{An optional location of the infector; if provided, it must be
a list with the following slots: 'location', 'dna'}

\item{generation_time}{The number of days separating the dates of infection
of the infector, and the infectee. Only needed if the infector is provided.}

\item{...}{further arguments passed to \code{\link{new_config}}}

\item{config}{optionally, a config returned by \code{\link{new_config}}}
}
\description{
This function creates the DNA sequence for a new case, with or without a known
infector. If the infector is missing (NULL, default), the case is imported according to
parameters in \code{config}.
}
\details{
We consider that all markers are biallelic, and store only
  mutations.
}
\examples{

## without initial location
x_dna <- new_dna()
x_dna

## drift from this sequence, different nb of generations
new_dna(x_dna, 1L) # 1 generation
new_dna(x_dna, 12L) # 12 generations

## 100 generations, other genetic params
new_dna(x_dna, 100, mutation_rate = 1e-6, genome_length = 2e6)

## illustrate reverse mutations
my_config <- new_config(mutation_rate = 0.1, genome_length = 10)
x <- integer(0)
set_seed(1)
x <- new_dna(x, 5); x
x <- new_dna(x, 5); x
x <- new_dna(x, 5); x
x <- new_dna(x, 5); x
x <- new_dna(x, 5); x
}
\seealso{
\code{\link{dist_dna}} to compute pairwise genetic distances from
  vectors of mutations.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
